[Parser] Fixed a datatype issue when shifting coordinates (Ringing Deeps) which caused random coords to remain inaccurate by up to 0.5 based on their starting value even after shifting

+++ .contrib/Source Code/Parser/Framework/Framework.Processing.cs
@@ -1083,11 +1083,11 @@ namespace ATT
                             LogWarn($"Non-Timelined data being coord-shifted due to {ToJSON(shiftInfo)}", data);
                         }
 
-                        if (coordObj[0].TryConvert(out long coordx))
+                        if (coordObj[0].TryConvert(out float coordx))
                         {
                             coordObj[0] = coordx + shiftInfo.X;
                         }
-                        if (coordObj[1].TryConvert(out long coordy))
+                        if (coordObj[1].TryConvert(out float coordy))
                         {
                             coordObj[1] = coordy + shiftInfo.Y;
                         }
